app/transaction-service/requirements.txt
Flask==2.3.2
flask-cors==4.0.0
psycopg2-binary==2.9.9
requests==2.31.0
gunicorn==21.2.0
